Add new-to-SoDA modal
The modal is triggered by a Learn More button in the first screenful of
the page, above the social media icons, which acts as our
call-to-action. Hopefully it catches people's attention. Frequently
asked questions are in the modal to address new members' concerns.

Though the button is generic, together with the text 'New to SoDA?' next
to the button, it is more indicative of its purpose.

The modal and button exist within a new LearnMore component.

FIXME: I've noticed that the modals don't capture keyboard focus. Also,
maybe the should have close or 'x' buttons.